WIP: Add For-like syntax to specify multiple rounds
Instead of manually writing the specification for each and every round
in a Cipher, we can instead specify multiple rounds of the same type
where only the index changes, with a new For-like syntax.

It works similar to macros in C, where the preprocessing would remove
all macros by substituition.

Currently the substituition works for Round names and Output values of
parts. Need to implement for input values.